Split, and move blazorwasm tests
