+++ src/mono/wasm/Wasm.Build.Tests/BuildTestBase.cs
@@ -786,6 +786,23 @@ namespace Wasm.Build.Tests
             return Path.Combine(dir!, "obj", config, targetFramework, "browser-wasm");
         }
 
+        protected string CreateProjectWithNativeReference(string id)
+        {
+            CreateBlazorWasmTemplateProject(id);
+
+            string extraItems = @$"
+                <PackageReference Include=""SkiaSharp"" Version=""2.88.1-preview.63"" />
+                <PackageReference Include=""SkiaSharp.NativeAssets.WebAssembly"" Version=""2.88.1-preview.63"" />
+
+                <NativeFileReference Include=""$(SkiaSharpStaticLibraryPath)\3.1.7\*.a"" />
+                <WasmFilesToIncludeInFileSystem Include=""{Path.Combine(BuildEnvironment.TestAssetsPath, "mono.png")}"" />
+            ";
+            string projectFile = Path.Combine(_projectDir!, $"{id}.csproj");
+            AddItemsPropertiesToProject(projectFile, extraItems: extraItems);
+
+            return projectFile;
+        }
+
         public static (int exitCode, string buildOutput) RunProcess(string path,
                                          ITestOutputHelper _testOutput,
                                          string args = "",
@@ -1032,4 +1049,6 @@ namespace Wasm.Build.Tests
         NativeFilesType ExpectedFileType,
         string TargetFramework = BuildTestBase.DefaultTargetFrameworkForBlazor
     );
+
+    public enum NativeFilesType { FromRuntimePack, Relinked, AOT };
 }
